significantly reduce code in vision file
