+++ fastchat/serve/gradio_block_arena_vision.py
@@ -2,125 +2,28 @@
 The gradio demo server for chatting with a single model.
 """
 
-import argparse
-from collections import defaultdict
-import datetime
-import json
 import os
-import random
-import time
-import uuid
 
 import gradio as gr
-import requests
 
-from fastchat.conversation import SeparatorStyle
-from fastchat.constants import (
-    LOGDIR,
-    WORKER_API_TIMEOUT,
-    ErrorCode,
-    MODERATION_MSG,
-    CONVERSATION_LIMIT_MSG,
-    SERVER_ERROR_MSG,
-    INACTIVE_MSG,
-    INPUT_CHAR_LEN_LIMIT,
-    CONVERSATION_TURN_LIMIT,
-    SESSION_EXPIRATION_TIME,
-)
 from fastchat.serve.gradio_web_server import (
-    State,
-    set_global_vars,
-    get_conv_log_filename,
-    get_model_list,
-    load_demo_single,
-    load_demo,
     upvote_last_response,
     downvote_last_response,
     flag_last_response,
-    post_process_code,
     get_model_description_md,
-    headers,
-    no_change_btn,
-    enable_btn,
-    disable_btn,
-    enable_moderation,
     acknowledgment_md,
-    block_css,
-    controller_url,
-    bot_response
+    bot_response,
+    add_text,
+    clear_history,
+    regenerate,
 )
-from fastchat.model.llava.constants import LLAVA_IMAGE_TOKEN
-from fastchat.model.model_adapter import get_conversation_template
 from fastchat.utils import (
     build_logger,
-    oai_moderation,
-    get_window_url_params_js,
-    get_window_url_params_with_tos_js,
-    parse_gradio_auth_creds,
 )
 
 logger = build_logger("gradio_web_server_multi", "gradio_web_server_multi.log")
 enable_moderation = False
 
-def regenerate(state, request: gr.Request):
-    logger.info(f"regenerate. ip: {request.client.host}")
-    state.conv.update_last_message(None)
-    return (state, state.to_gradio_chatbot(), "", None) + (disable_btn,) * 5
-
-
-def clear_history(request: gr.Request):
-    logger.info(f"clear_history. ip: {request.client.host}")
-    state = None
-    return (state, [], "", None) + (disable_btn,) * 5
-
-
-def add_text(state, model_selector, text, image, request: gr.Request):
-    ip = request.client.host
-    logger.info(f"add_text. ip: {ip}. len: {len(text)}")
-
-    if state is None:
-        state = State(model_selector)
-
-    if len(text) <= 0:
-        state.skip_next = True
-        return (state, state.to_gradio_chatbot(), "", None) + (no_change_btn,) * 5
-
-    if enable_moderation:
-        flagged = oai_moderation(text)
-        if flagged:
-            logger.info(f"violate moderation. ip: {request.client.host}. text: {text}")
-            state.skip_next = True
-            return (state, state.to_gradio_chatbot(), MODERATION_MSG, None) + (
-                no_change_btn,
-            ) * 5
-
-    if image is not None and len(state.conv.get_images()) > 0:
-        # reset convo with new image
-        state.conv = get_conversation_template(state.model_name)
-
-    conv = state.conv
-    if (len(conv.messages) - conv.offset) // 2 >= CONVERSATION_TURN_LIMIT:
-        logger.info(f"conversation turn limit. ip: {request.client.host}. text: {text}")
-        state.skip_next = True
-        return (state, state.to_gradio_chatbot(), CONVERSATION_LIMIT_MSG, None) + (
-            no_change_btn,
-        ) * 5
-
-    text = text[:INPUT_CHAR_LEN_LIMIT]  # Hard cut-off
-
-    if image is not None:
-        if "llava" in state.model_name.lower():
-            if LLAVA_IMAGE_TOKEN not in text:
-                text = text + "\n" + LLAVA_IMAGE_TOKEN
-
-        text = (
-            text,
-            image,
-        )
-
-    conv.append_message(conv.roles[0], text)
-    conv.append_message(conv.roles[1], None)
-    return (state, state.to_gradio_chatbot(), "", None) + (disable_btn,) * 5
 
 def build_single_vision_language_model_ui(models, add_promotion_links=False):
     promotion = (
